Update patch set 1
Patch Set 1:

This has come up a number of times now in different contexts. I think we probably need to bite the bullet and write a configure test that verifies that the pthreads library is linkable for the symbols we want, rather than only that the header exists.

That said, I'm ok with defining that macro for now too. If we ever get a fully async API we may need to revisit.

Patch-set: 1
Label: Code-Review=0
Label: Commit-Queue=0