Update patch set 3
Patch Set 3: (1 inline comment)

As a general note, these functions need to be replaced with ones that take the height as a parameter for support of 4:2:2. The 1:4 and 4:1 rectangles happen in that case, and it'd be better from a code size perspective. It doesn't have to happen now, but if you're rewriting everything already, maybe it's easy.

Patch-set: 3
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0