Test a new template matching intra prediction mode.
This patch conducts experiments of an additional intra prediction mode
based on template matching. For each target block to be coded, its causal
borders (left and above) are used as its L-shaped template. The most similar
L-shaped area to this template is found among the already coded pixels
within the frame. The block corresponding to this L-shaped area is then
used as the prediction for the target block. Since the decoder can repeat
the searching process, no motion / displacement vectors are needed to be sent.
This mode is called "BM" (Block Matching) mode in the patch. ADST transform is
used to encode the residue signals under BM mode. This patch only considers
8X8 partitions.

The gain of this new BM mode depends on the sequcne contents, whether or not
they contain many repeated patterns. For example, it works well on
foreman (2~3.5%), ice (2.5~4%), and carphone (0.8~1.6%). A presentation
slides of this patch can be found at:
http://yon.se/AGVHC2

Change-Id: Ia4072d17355385d846ca40bfecc60463314a5005