Tx size selection based on model
Modifies the modeling fucntion to account for tx size. Uses that
to add some speed features. Specifically, intead of using a
flag use_largest_txfm as a speed feature an enum tx_size_search_method
is used, of which two of the types are USE_FULL_RD and
USE_LARGEST_TXFM. Two other new types is added: USE_MODEL_FOR_32INTER
and USE_MODEL_FOR_3216INTER. These use the model instead of full-rd
for etimating the cost of the 32x32 and 16x16 transforms.

Another change is that the framework for deciding transform type
is simplified to use a heuristic count based method rather than
an rd based method using txfm_cache. In practice the new method
is found to work just as well - with derf only -0.001 down.
The new method is more compatible with the new framework where
certain rd costs are based on full rd and certain others are
based on modeled rd. In this patch the existing rd based method
is kept for use in the USE_FULL_RD mode. In the other modes,
the count based method is used.

Some optimizations of the scales for the different tx_types in the
modeling fucntion is pending.

Results on derf:
USE_FULL_RD: -0.01%, 0 speedup
USE_MODEL_FOR_32INTER: -0.20%, 5% speedup
USE_MODEL_FOR_1632INTER: -0.87%, 9% speedup
USE_LARGEST_TXFM (existing use_largest_txfm flag): -1.10%, 10% speedup

Change-Id: I4d60a5f9ce78fbc90cddf2f97ed91d8bc0d4f936