Quantize (64-bit only, for now) SSSE3 SIMD.
Total encoding time for first 50 frames of bus (speed 0) @ 1500kbps
goes 2min43.9 to 2min20.8, i.e. a 16.4% overall speedup. I'll try to
modify the SIMD to be 32-bit compatible over the next few days; the
GPR usage is 32bit compliant, it's only the XMM register usage that
isn't (it uses 15, but 32bit only has 8, so we need some stack space).
Alternatively, I can mark the function to be compiled only on 64bit.
Also, we may want a SSE2 version as fallback for systems that don't
have SSSE3.

Also, the 32x32 quantize still uses C code. The SIMD needs some minor
modifications and then it can be used for the 32x32 DCT as well.

Change-Id: I2df53770c2e850813ffa713e1a91b45b0082b904