Implements several heuristics to prune mode search
Skips mode searches for intra and compound inter modes depending
on the best mode so far and the reference frames. Also makes
an initial implementation of early termination of mode search,
which is to be refined later. The various heuristics to be used
are selected by bits from a flag.

Results: pending for the various combinations.
Speedup of up to 20% is observed.

Change-Id: I1a736cd29b36325489e7af9f32698d6394b2c495