Implements several heuristics to prune mode search
Skips mode searches for intra and compound inter modes depending
on the best mode so far and the reference frames. Also makes
an initial implementation of early termination of mode search,
which is to be refined later. The various heuristics to be used
are selected by bits from a flag. The previous direction based
intra mode search pruning is also absorbed in this framework.

Specifically the flags and their impact are:

1) FLAG_SKIP_INTRA_BESTINTER (skip intra mode search for oblique
directional modes and TM_PRED if the best so far is
an inter mode)

2) FLAG_SKIP_INTRA_DIRMISMATCH (skip D27, D63, D117 and D153
mode search if the best so far is not one of the closest
hor/vert/diagonal directions.

3) FLAG_SKIP_COMP_BESTINTRA (skip compound prediction mode
search if the best so far is an intra mode)
derfraw300: (results pending)

4) FLAG_SKIP_COMP_REFMISMATCH (skip compound prediction search
if the best single ref inter mode does not have the same ref
as one of the two references being tested in the compound mode)
derfraw300: (results pending)

5) FLAG_EARLY_TERMINATE (terminate mode search based on
distortion achieved so far with respect to a metric based on
the qpstep)
derfraw300: (results pending)

Results:
pending

Change-Id: I1a736cd29b36325489e7af9f32698d6394b2c495