Eliminated prev_mip memsets/memcpys in encoder
This patch is in experimental but was not merged into master.

This patch swaps ptrs instead of copying and uses the
last show_frame flag instead of setting the entire buffer
to zero.  On the decode side, forced prev_mi to always point
to a valid mode info buffer.  Required by the next frame.

Change-Id: Ia0950466c8ba301a2a5bf917ff3d07bc1a2c2311