Add encoding option --static-thresh
This option exists in VP8, and it was rewritten in VP9 to support
skipping on different partition sizes. After prediction is done,
we can check if the residuals in the partition block will be all
quantized to 0. If this is true, the skip flag is set, and only
prediction data are needed in reconstruction. Based on DCT's energy
conservation property, the skipping check can be estimated in
spatial domain.

The prediction error is calculated and compared to a threshold.
The threshold is determined by the dequantizer value, and also
adjusted by partition sizes. To be precise, the DC and AC parts
for Y, U, and V planes are checked to decide skipping or not.

Test on derf set showed that:
when static-threshold = 1, psnr loss is 0.666%;
when static-threshold = 500, psnr loss is 1.162%;

For different clips, the speedup range is between several
percentage and up to 20%.

Change-Id: I857031ceb466ff314ab580ac5ec5d18542203c53