Reworked the auto_mv_step_size speed feature
This patch modifies the auto_mv_step_size speed feature to
use the maximum magnitude mv from the last inter frame, rather
than the last arf frame. The bounds therefore are slightly
tighter. The feature is made a speed 1 feature.

Results (when this feature is turned on for speed 0):
derfraw300: -0.02% psnr, about +1% speedup.

Change-Id: If492797a61b0b4b3e58c0b8f86afb880165fc9f6