Use hex search conditionally
Adds a speed feature to use faster hex search conditionally
based on max mv expected.

Some changes are also made to the hex search function to
allow multi-step search.

With the feature on at speed 0, results are -1% or derf, at
a speedup of 4-5%. The feature is made part of the speed 2
and higher setting, since the trade-off is not ideal at speed
1.

Change-Id: I02a7ef5193f762601e0994e2c99399a3535a43d2