Auto min and max partition size experiment.
Speed feature experiment to set an upper and lower
partition size limit based on what has been seen
in spatial neighbors.

This seems to gives quite reasonable speed gains in local
(10-25%) and when used with speed 0 the losses are small
(0.25% derf, 0.35% stdhd). However, for now I am only
enabling it on speed 1 as there may be clashes with the existing
temporal partition selection in speed 2.

I think this spatial method could be combined with data from either
the last frame or a variance method (or both) to refine the
choice of minimum and maximum partition size. I.e. consider the
min and max from spatial and temporal neighbors and the variance
recommendation.

Change-Id: I1b96bf8b84368d6aad0c7aa600fe141b4f07435f