Add variance based mode/skipping
Adds a speed feature to skip all intra modes other than
DC_PRED/TM_PRED if the source variance is small. This
feature is made part of speed 1 and up.

Results on derf300: psnr -0.06%, speedup about 1-2%

Also uses the source variance to fine-tune the early
termination criteria when FLAG_EARLY_TERMINATE is on.
This feature is made part of speed 2 and up.

Results on derf300: psnr -0.52%, speedup about 5-7%

Change-Id: I59e38aa836557cfa5405ae706fc64815cbfe4232