Update patch set 1
Patch Set 1:

These are used in decoding also, so yes they should obviously be enabled when possible. Alternatively, someone should rewrite all this in intrinsics and verify it's not slower.

x86inc.asm works when:
(A) we're 64bit
(B) we're 32bit, but pic is disabled
(C) we're 32bit, pic is enabled, but the final binary is linked with -mdynamic-no-pic (or a non-gcc equivalent).

x86inc.asm doesn't work when:
(D) we're 32bit, pic is enabled, and the final binary does not use -mdynamic-no-pic or a non-gcc equivalent during linking

My understanding is that Chrome on win32 does (C) (i.e. the windows/msvc linker does gcc's -mdynamic-no-pic behaviour by default), so there's definitely an argument to be made for keeping these assembly optimizations.

Patch-set: 1