Affine Motion Prediction Experiment
This patch tests an affine motion prediction scheme. An affine
inter prediction mode is added where we fit an affine transform
between the target block and the reference block. Several small
blocks that have been previously encoded are used to find some
anchor point correpondences between the current frame and the
reference frame, based on which the affine coefficients are
estimated. The decoder can repeat the coefficient estimation
process, so no overhead is encoded. This experiment is a proof
of concept. The partition size is fixed (either 8x8 or 16x16);
the reference frame is forced to be last frame only. On derf,
the gain is 1.14% for 16x16 blocks and 0.93% for 8x8 blocks.

Change-Id: Ic8aaa0728a43936cd4c6e1ed590e01ba8f0fbf5b