Update patch set 1
Patch Set 1: (1 inline comment)

The build error is coming from a case we don't usually see where the email you're using with Gerrit doesn't match the email on the commit message. We use gerrit's email because it's available as an environment variable and we haven't checked out the source at this point, so we can't check the email on the commit. It's a pretty easy change in Jenkins to do the checkout, and I can walk Johann or jzern@ through it (or they can just look at 'vpnext per-commit trigger' for an example, I think we made that change there already)

Patch-set: 1
Label: Code-Review=0
Label: Commit-Queue=0