Masked Compound Inter Prediction
The masked compound motion compensation has mask types separating a
block into wedges at specific angles and offsets. The mask is used to
weight pixels from the first and second predictors to obtain the final
predictor. The weighting is smooth near the partition boundaries but
becomes a selecton farther away.

Change-Id: I59eff5c59ea07ac13611a7d4575981ba83b6818b