Clean ups of the subpel search functions
Removes some unused code and speed features, and organizes the
interfaces for fractional mv step functions for use in new speed
features to come.

In the process a new speed feature - number of iterations per
step during the subpel search - is exposed.

No bitstream or output change when this parameter is set as the
original value of 3.

Change-Id: I2eba8a21f6461be8caf56af04a5337257a5693a8