Update patch set 1
Patch Set 1:

This is first time I'm preparing a release, so some mistakes are to be expected.

Also, I'm not _that_ familiar with libvpx and even media codecs, so some suggestions to summarize/reword the changelog would be very welcome.

I also may have missed some updates that are in fact important but just didn't look so to me.

Patch-set: 1
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0