Make "good" quality 2-pass vpxenc encoding default
Currently, the best quality mode in VP9 is not very well developed,
and unnecessarily makes the encode too slow. Hence the command line
default is changed to "good" quality. Besides, the number of passes
default is changed to 2 passes as well, since 1-pass encoding is
not very efficient in VP9.

Note that this also changes the defaults for VP8.

Change-Id: Ib23c1a05ae2f36ee076c0e34403efbda518c5066