Update patch set 5
Patch Set 5: I would prefer that you didn't submit this

It seems to me that we don't want any of our code to be using this.  It might make the valgrind error go away but there's no code in libvpx that should use the border value 0 that this does.

Patch-set: 5
Label: Code-Review=-1
Label: Commit-Queue=0
Label: Verified=0