Update patch set 1
Patch Set 1: Abandoned

After chatting with Paul, it was determined that this check is redundant.  This function is called on sb64 mode info boundaries, so if left[-1] is outside the visible area, left[-8] will be as well.  I was thinking it could be called on 8x8 mode info boundaries.

Patch-set: 1