Modified mode skip functionality.
A previous speed feature skipped modes not used in earlier
partitions but this not longer worked as intended following
changes to the partition coding order and in conjunction
with some other speed features.

This modified mode skip feature sets a mask after the first X
modes have been tested in each partition depending on the
reference frame of the current best case.

This patch also makes some changes to the order modes are
to fit better with this skip functionality.

Initial testing suggests speed and rd hit count improvements
of 10-20% with a modest quality hit. Full quality metrics to follow.

Change-Id: Idd8efa656cbc0c28f06d09690984c1f18b1115e1