Modified resize unit test to output test vector
Modified the resize unit test so that it optionally
writes the encoded bitstream to file. The macro
WRITE_COMPRESSED_STREAM should be set to 1 to enable
output of the test bitstream; it is set to 0 by default.

In addition, modified libs.mk to make the
LIBVPX_TEST_DATA target dependent on test-data.sha1 so
that if modified test vectors are created with an
existing filename, they are automatically re-copied
when the unit tests are run. (Thanks to jzern@ for
figuring that one out!).

Change-Id: I7d436b1942f935da97db6d84574a98d379f57fb1