Added encoder control to set absolute frame size
Added an encoder control to enable the frame size to be
set to a defined (width, height). Previously, the frame
size could only be set either to a defined fraction of
the current frame size, or the width and height could be
set independently (requiring two reconfigurations rather
than one for the underlying buffers).

In addition, I extended the resize unit test to test
this new resizing mode.

Change-Id: I496730e205a6266c9beaaccc23ecf47fb3ce5a2a