Improve variance-based partitioning
Before this, we only stopped splitting if the curent block variance was
low enough. Now we also stop splitting if its variance is close enough
to the smallest variance of its 4 subblocks.

The various constants used could probably be tuned to produce better
results, and the logarithmic calculations could be eliminated if needed.

Overall PSNR improvements at speed 3:
derf: +7.719%
stdhd: +93.357%

Change-Id: I5def3253a2f8d99ad78fd3c6c74111b3cb0c8f48