Update patch set 1
Patch Set 1:

I should have caught this when you first submitted it. You're storing the registers to the stack, then adjusting the stack when you return to the calling function, and getting lucky that the values are still available for the restore.

A proper implementation saves the registers into a variable.

https://code.google.com/p/webm/source/browse/vp8/common/arm/neon/save_reg_neon.asm?repo=libvpx

Patch-set: 1