BITSTREAM - FIX "update_map" SEMANTICS BROKEN IN 398ddafb629b7f49cf255bf09d3e38b4abd0bb95
"update_map" set to 0 means that the segmentation map should be the same
as the map of the last frame with update_map set to 1 (except if a key
frame is in between, in that case all the segments are set to 0).
398ddafb629b7f49cf255bf09d3e38b4abd0bb95 accidentally changed this so
that the map is still implicitely updated at every frame because of
partitioning changes. This breaks the bitstream and doesn't match what
the encoder does.

Here's a simple example demonstrating the problem:
* Block 0 in Frame 0:
  - update_map = 1
* Block 0 in Frame 1:
  - update_map = 0
  - different partitioning
* Block 0 in Frame 2:
  - update_map = 0
  - same partitioning as in Frame 0

* Incorrect behavior introduced by 398ddafb629b7f49cf255bf09d3e38b4abd0bb95:
+-+-+       +-+-+       +-+-+
|0|1|       |   |       |0|0|
+-+-+ ----> + 0 + ----> +-+-+
|1|0|       |   |       |0|0|
+-+-+       +-+-+       +-+-+

* Correct behaviour:
+-+-+       +-+-+       +-+-+
|0|1|       |   |       |0|1|
+-+-+ ----> + 0 + ----> +-+-+
|1|0|       |   |       |1|0|
+-+-+       +-+-+       +-+-+

Change-Id: Ibeb623261c3d6b8b7667af2c21315e8f893724c4