Update patch set 6
Patch Set 6:

Awesome! The main reason I didn't push it was because we lacked any unit test running to make sure that it works.   My guess is that we could reuse one of the vp8 tests and prune out tools,  and get it to work.

Patch-set: 6
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0